🔀 Merge Pull Request #29

+++ src/main/java/com/salomovs/bookiero/model/entity/Book.java
@@ -42,7 +42,7 @@ public class Book {
   private String editor;
 
   @Column(name="in_stock_amount", nullable=false)
-  private Integer inStockAmount;
+  private Long inStockAmount;
 
   @Column(name="book_cover", nullable=true)
   private String bookCover;
